+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+1.2.0 / WIP
+------------------
+
+- Fixed wrong approximation for very small cubic curves: when diff between
+  start and end of a cubic curve falls within `errorBound`, in earlier
+  versions any quadratic curve was considered to be fitting.
+  Error estimation algorithm is changed: previously it calculated distance
+  from points on cubic curve to quadratic, now it also calculates the other
+  way around. Distance calculation is changed from exact solution with
+  cubic solver to linear segment approximations of both curves.
+
+
 1.1.1 / 2017-05-08
 ------------------